requirements.txt
asgiref==3.7.2
Django==5.0.1
django-enumchoicefield==3.0.1
django-extensions==3.2.3
djangorestframework==3.14.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
boto3==1.34.46
django-storages==1.14.2